+++ seccardgame/src/cli/prompts.rs
@@ -4,21 +4,18 @@ pub fn prompt<T: std::str::FromStr + Default>(
     prompt_msg: &str,
     validator: Option<Box<dyn Fn(&T) -> Result<(), String>>>,
 ) -> T
-    where
-        <T as std::str::FromStr>::Err: std::fmt::Debug + ToString, T: Clone + ToString
+where
+    <T as std::str::FromStr>::Err: std::fmt::Debug + ToString,
+    T: Clone + ToString,
 {
     let mut input = Input::<T>::new().with_prompt(prompt_msg);
     if let Some(validator) = validator {
         input = input.validate_with(validator);
     }
     input.interact_text().unwrap()
-
 }
 
-pub fn prompt_allow_empty(
-    prompt_msg: &str,
-) -> String
-{
+pub fn prompt_allow_empty(prompt_msg: &str) -> String {
     let default = "";
 
     Input::<String>::new()