+++ Changes.md
@@ -1,4 +1,11 @@
-# Changes
+#Changes
+
+## upcoming release
+
+* introduces flag to supply config file
+* adds init command
+  * clones cards repository
+  * creates config file
 
 ## 0.1.0-alpha